+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## [UNRELEASED]
+## [0.16.0]
 ### Fixed/Changed
  - Added `DisplayNameOptions` to methods for language, script, and region names.
  - Lazy initialize some internal instances.
@@ -169,6 +169,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 #### Fixed/Changed
 - Converted `CLDR` to an interface
 
+[0.16.0]: https://github.com/phensley/cldr-engine/compare/v0.15.3...v0.16.0
 [0.15.3]: https://github.com/phensley/cldr-engine/compare/v0.15.2...v0.15.3
 [0.15.2]: https://github.com/phensley/cldr-engine/compare/v0.15.1...v0.15.2
 [0.15.1]: https://github.com/phensley/cldr-engine/compare/v0.15.0...v0.15.1